implements dependecy inversion on management of migrations.
