+++ tests/seed.go
@@ -3,7 +3,8 @@ package main
 import (
 	"fmt"
 
-	pkg "github.com/Vinicius-Santos-da-Silva/mongo-migrate/pkg"
+	repository "github.com/Vinicius-Santos-da-Silva/mongo-migrate/pkg/database_adapter"
+	pkg "github.com/Vinicius-Santos-da-Silva/mongo-migrate/pkg/migrate"
 	seeds "github.com/Vinicius-Santos-da-Silva/mongo-migrate/tests/seed"
 )
 
@@ -16,6 +17,9 @@ func main() {
 	if err != nil {
 		panic(err)
 	}
+
+	repo := repository.NewMigrationRepositoryMongo(database)
+	pkg.SetRepository(repo)
 
 	pkg.Register(seeds.NewAddMyIndex(database))
 	pkg.Register(seeds.NewAddMyIndex(database))
